ecommerce_analytics/pyproject.toml
[project]
name = "ecommerce-analytics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.11.9"
dependencies = [
    "apache-airflow==2.8.1",
    "astro>=0.2.7",
    "dbt-core==1.9.0",
    "dbt-duckdb>=1.9.4",
    "dbt-postgres==1.9.0",
    "faker>=37.4.2",
    "pandas>=2.3.1",
    "python-dotenv>=1.1.1",
]
